FaceBot/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
-e git+https://github.com/mobolic/facebook-sdk.git@3fa89fec6a20dd070ccf57968c6f89256f237f54#egg=facebook_sdk
idna==3.4
python-dotenv==0.21.0
requests==2.28.1
schedule==1.1.0
urllib3==1.26.12
